Fixed click outside
